Extract logic so test file doesnt register tsx
